requirements.txt
alembic==1.7.5
antiorm==1.2.1
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
db==0.1.1
db-sqlite3==0.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.1.1
zipp==3.8.0
